Extract type data to separate module
